Improved image loading, Removed JFrame limitation
Added image loading to FileIOHelper with  type optimizations.
JGame class originally forced the use of JFrame, now it's independent of
its container and can be used with any frame/applet.

+++ src/jgame/Updatable.java
@@ -1,5 +1,5 @@
 package jgame;
 
 public interface Updatable {
-	public void update(float dt);
+	public void update(double dt);
 }
